[renderer] use accessflags2 instead of accessflags
accessflags2 gives us better granularity over access types, we want to
use this to keep track of synchronisation state for our resources
