+++ modules/le_renderer/le_command_buffer_encoder.cpp
@@ -139,7 +139,7 @@ static void cbe_dispatch( le_command_buffer_encoder_o* self, uint32_t groupCount
 static void cbe_buffer_memory_barrier( le_command_buffer_encoder_o*   self,
                                        le::PipelineStageFlags2 const& srcStageMask,
                                        le::PipelineStageFlags2 const& dstStageMask,
-                                       le::AccessFlags const&         dstAccessMask,
+                                       le::AccessFlags2 const&        dstAccessMask,
                                        le_buf_resource_handle const&  buffer,
                                        uint64_t const&                offset,
                                        uint64_t const&                range ) {
