+++ modules/le_renderer/private/generate_enums.sh
@@ -2,7 +2,7 @@
 
 ../../../scripts/codegen/gen_le_enums.py <<- "EOF" | clang-format > le_vk_enums.inl
 # Vulkan enum name,         generate to_string()
-VkAccessFlagBits
+VkAccessFlagBits2
 VkAttachmentLoadOp, yes
 VkAttachmentStoreOp,        yes
 VkBlendFactor, yes
